changed Training repair to different format
